refactor colors
