+++ src/styles/colorData.ts
@@ -1,8 +1,19 @@
-import { css } from '@emotion/react';
-import type { Extras } from '../logic/types';
-import { neutral, space, from, until, body, headline } from '@guardian/source-foundations';
+import { brandAlt, neutral } from '@guardian/source-foundations';
 
-const imgHeight = '280';
+const buttonStyles = {
+    textPrimary: neutral[7],
+    backgroundPrimary: brandAlt[400],
+    backgroundPrimaryHover: brandAlt[300],
+    textSecondary: neutral[7],
+    backgroundSecondary: neutral[93],
+    backgroundSecondaryHover: neutral[86],
+    borderSecondary: neutral[86],
+};
+
+export const contributionsTheme = {
+    button: buttonStyles,
+    link: buttonStyles,
+};
 
 export interface StyleData {
     styleBackground: string;
@@ -22,7 +33,7 @@ export interface StyleData {
     styleCloseHover: string;
 }
 
-const colorStringStyles = [
+export const colorStringStyles = [
     'styleBackground',
     'styleHeader',
     'styleBody',
@@ -40,292 +51,4 @@ const colorStringStyles = [
     'styleCloseHover',
 ];
 
-type Styles = keyof StyleData;
-
-export const selfServeStyles = (userVals: Extras, defaults: StyleData) => {
-    const style: StyleData = Object.assign({}, defaults);
-    const defKeys: Styles[] = Object.keys(defaults) as Styles[];
-    const regex = new RegExp(/^#([A-Fa-f0-9]{6})$/);
-
-    defKeys.forEach((key) => {
-        const userVal = userVals[key];
-
-        // If user val is undefined, or an empty string, use default val
-        if (userVal == null || !userVal.length) {
-            return;
-        }
-
-        // Protect against CSS injection
-        const item = userVal.split(';')[0].trim();
-
-        // Protect against null or empty user strings
-        if (item == null || !item.length) {
-            return;
-        }
-
-        // Check for legitimate CSS color string values
-        // - we only support `#abcdef` color format
-        if (colorStringStyles.includes(key) && regex.test(item)) {
-            style[key] = item;
-        }
-    });
-
-    return {
-        wrapper: css`
-            box-sizing: border-box;
-            display: flex;
-            flex-direction: column;
-            justify-content: center;
-            align-items: center;
-            width: 100%;
-            background-color: ${style.styleBackground};
-            color: ${neutral[20]};
-
-            html {
-                box-sizing: border-box;
-            }
-
-            *,
-            *:before,
-            *:after {
-                box-sizing: inherit;
-            }
-        `,
-
-        contentContainer: css`
-            align-items: stretch;
-            display: flex;
-            flex-direction: column;
-            position: relative;
-            margin: 0 auto;
-            width: 100%;
-            max-width: 740px;
-
-            ${from.desktop} {
-                max-width: 980px;
-                flex-direction: row;
-                min-height: ${imgHeight}px;
-            }
-
-            ${from.leftCol} {
-                max-width: 1140px;
-            }
-
-            ${from.wide} {
-                max-width: 1300px;
-            }
-        `,
-
-        topLeftComponent: css`
-            width: 93%;
-            padding: ${space[4]}px;
-            min-height: ${imgHeight}px;
-            ${from.desktop} {
-                width: 60%;
-            }
-        `,
-
-        bottomRightComponent: css`
-            display: flex;
-            justify-content: center;
-            width: 100%;
-            max-height: 100%;
-
-            ${from.desktop} {
-                align-self: flex-end;
-                padding-right: ${space[4]}px;
-                max-width: 45%;
-                justify-content: flex-end;
-            }
-
-            ${from.leftCol} {
-                padding-right: 0;
-                justify-content: space-between;
-            }
-
-            ${from.wide} {
-                max-width: 48%;
-            }
-        `,
-
-        heading: css`
-            ${headline.small({ fontWeight: 'bold' })};
-            margin: 0;
-            max-width: 100%;
-            color: ${style.styleHeader};
-
-            ${from.mobileLandscape} {
-                ${headline.small({ fontWeight: 'bold' })};
-            }
-
-            ${from.tablet} {
-                max-width: 100%;
-            }
-        `,
-
-        paragraph: css`
-            ${body.medium()}
-            line-height: 135%;
-            margin: ${space[5]}px 0 ${space[5]}px;
-            max-width: 100%;
-            color: ${style.styleBody};
-
-            ${from.phablet} {
-                max-width: 90%;
-            }
-
-            ${from.tablet} {
-                max-width: 100%;
-            }
-
-            ${from.desktop} {
-                font-size: 20px;
-                margin: ${space[3]}px 0 ${space[4]}px;
-                max-width: 42rem;
-            }
-
-            ${from.leftCol} {
-                max-width: 37rem;
-            }
-
-            ${from.wide} {
-                max-width: 42rem;
-            }
-        `,
-
-        secondParagraph: css`
-            ${from.desktop} {
-                font-size: 18px;
-            }
-        `,
-
-        highlight: css`
-            font-weight: 700;
-            margin-top: ${space[5]}px;
-            margin-right: ${space[3]}px;
-            display: inline-block;
-            color: ${style.styleHighlight};
-            background-color: ${style.styleHighlightBackground};
-        `,
-
-        smallRightSpacer: css`
-            margin-right: ${space[3]}px;
-        `,
-
-        primaryButtonWrapper: css`
-            margin: ${space[4]}px ${space[2]}px ${space[1]}px 0;
-            display: flex;
-            flex-wrap: wrap;
-            align-items: center;
-
-            &.hidden {
-                display: none;
-            }
-        `,
-
-        primaryButton: css`
-            margin-right: ${space[3]}px;
-            color: ${style.styleButton};
-            background-color: ${style.styleButtonBackground};
-            &:hover {
-                background-color: ${style.styleButtonHover};
-            }
-        `,
-
-        iconPanel: css`
-            ${from.desktop} {
-                display: flex;
-                flex-direction: column;
-                justify-content: space-between;
-                align-items: flex-end;
-                height: 100%;
-                padding: ${space[4]}px 0;
-                margin-left: ${space[4]}px;
-            }
-        `,
-
-        closeButton: css`
-            display: flex;
-            align-items: center;
-            justify-content: center;
-            padding: 0;
-            position: absolute;
-            top: 15px;
-            right: 10px;
-            border: 1px solid ${style.styleClose};
-            background-color: ${style.styleCloseBackground};
-            &:hover {
-                background-color: ${style.styleCloseHover};
-            }
-            & svg {
-                fill: ${style.styleClose};
-            }
-        `,
-
-        image: css`
-            max-width: 100%;
-            max-height: 300px;
-            display: flex;
-            justify-content: center;
-            align-items: flex-end;
-            margin-top: ${space[2]}px;
-
-            img {
-                max-width: 100%;
-                width: 100%;
-                height: 100%;
-                object-fit: contain;
-            }
-
-            ${until.desktop} {
-                display: none;
-            }
-
-            ${from.wide} {
-                margin-right: 100px;
-            }
-        `,
-        centeredBottomRightComponent: css`
-            display: flex;
-            justify-content: center;
-            width: 100%;
-            max-height: 100%;
-
-            ${from.desktop} {
-                padding-right: ${space[4]}px;
-                max-width: 45%;
-            }
-
-            ${from.leftCol} {
-                padding-right: 0;
-            }
-
-            ${from.wide} {
-                max-width: 48%;
-            }
-        `,
-        centeredImage: css`
-            max-width: 100%;
-            max-height: 300px;
-            display: flex;
-            justify-content: center;
-            align-items: center;
-            margin-top: ${space[2]}px;
-
-            img {
-                max-width: 100%;
-                width: 100%;
-                height: 100%;
-                object-fit: contain;
-            }
-
-            ${until.desktop} {
-                display: none;
-            }
-
-            ${from.wide} {
-                margin-right: 100px;
-            }
-        `,
-    };
-};
+export type Styles = keyof StyleData;
